+++ tests/MediaTypeTest.php
@@ -130,10 +130,12 @@ class MediaTypeTest extends TestCase
                 Set::strings()->exclude(static fn($type) => (bool) \preg_match('~^[A-Za-z0-9][A-Za-z0-9!#$&^_.-]{0,126}$~', $type)),
             )
             ->prove(function($type) {
-                $this->expectException(\DomainException::class);
-                $this->expectExceptionMessage($type);
-
-                MediaType::from(TopLevel::application, $type);
+                $this
+                    ->assert()
+                    ->throws(
+                        static fn() => MediaType::from(TopLevel::application, $type),
+                        \DomainException::class,
+                    );
             });
     }
 
@@ -147,7 +149,7 @@ class MediaTypeTest extends TestCase
             )
             ->prove(function($suffix) {
                 try {
-                    MediaType::from(TopLevel::application, 'json', $suffix);
+                    $_ = MediaType::from(TopLevel::application, 'json', $suffix);
                     $this->fail('it should throw');
                 } catch (\DomainException $e) {
                     $this->assertSame($suffix, $e->getMessage());